+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
+- MongoDB instrumentation is updated to comply with v1.39.0 Semantic Convention
+
 #### Dependency updates
 
 ### Deprecated
@@ -55,7 +57,6 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
-- Update `db.system` attribute to `db.system.name` in MongoDB instrumentation.
 - `-NoReset` parameter in `Unregister-OpenTelemetryForIIS` changed from
   `[bool]` to `[switch]`. Use `-NoReset` instead of `-NoReset $true`.
   Note: `-NoReset $true` still works, but `-NoReset $false` will not do what you